pyproject.toml
[tool.poetry]
name = "aio-usb-hotplug"
version = "3.0.1"
description = "Asynchronous generators yielding detected hotplug events on the USB buses"
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ntamas/aio-usb-hotplug/"
repository = "https://github.com/ntamas/aio-usb-hotplug/"
authors = ["Tamas Nepusz <tamas@collmot.com>"]
packages = [
    { include = "aio_usb_hotplug", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
anyio = "^3.0.0"
pyudev = { version = "^0.22.0", platform = "linux" }
pyusb = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
coverage = {extras = ["toml"], version = "^5.2"}
pytest-cov = "^2.10.0"
curio = "^1.4"
trio = "^0.17.0"

[tool.poetry.scripts]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
